py/requirements.txt
Pillow
numpy
leb128